# OTLP: Use PrometheusConverter directly