Optimize by using `prometheusremotewrite.PrometheusConverter` directly in the OTLP endpoint handler, instead of going via the wrapping `prometheusremotewrite.FromMetrics` function. Follow-up to #13991.

Also add benchmark for `PrometheusConverter.FromMetrics`.

#### Benchmarks

<details>
<summary>Benchmark stats vs main branch</summary>

```console
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite
                                                                                                                                           │ frommetrics.txt │ prometheusconverter-frommetrics.txt │
                                                                                                                                           │     sec/op      │    sec/op     vs base               │
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12            1.802m ±  4%   1.672m ±  3%   -7.21% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12            3.195m ±  3%   3.074m ±  1%   -3.79% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12           4.305m ±  2%   4.152m ± 10%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12           10.56m ±  1%   10.55m ±  2%        ~ (p=0.818 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12           12.17m ±  3%   12.13m ±  1%        ~ (p=0.589 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12          13.38m ±  3%   13.19m ±  1%        ~ (p=0.093 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12            5.859m ±  1%   5.398m ±  2%   -7.86% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12            7.857m ±  4%   7.623m ± 11%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12           9.181m ± 12%   8.843m ±  1%   -3.68% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12           18.33m ±  1%   18.12m ±  4%        ~ (p=0.394 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12           20.30m ±  1%   19.90m ±  1%   -1.96% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12          21.75m ±  2%   21.90m ±  3%        ~ (p=0.485 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12         7.368m ±  3%   6.865m ±  3%   -6.83% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12         10.34m ±  6%   10.26m ±  3%        ~ (p=0.394 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12        13.08m ±  2%   12.88m ±  1%   -1.55% (p=0.015 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12        29.72m ±  3%   29.21m ±  4%   -1.73% (p=0.026 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12        32.54m ±  1%   32.48m ±  3%        ~ (p=0.818 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12       35.45m ±  4%   34.95m ±  8%        ~ (p=0.699 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12               341.0n ±  0%   299.5n ±  1%  -12.18% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12               342.6n ±  1%   301.8n ±  1%  -11.91% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12              342.8n ±  3%   299.9n ±  6%  -12.53% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12              343.4n ±  3%   297.2n ±  2%  -13.45% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12              344.5n ±  4%   302.1n ±  5%  -12.31% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12             349.0n ±  4%   299.9n ±  1%  -14.07% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12            1.873m ±  2%   1.705m ±  1%   -8.98% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12            3.231m ±  5%   3.094m ±  2%   -4.25% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12           4.331m ±  7%   4.201m ±  6%   -2.99% (p=0.041 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12           10.84m ±  6%   10.59m ±  7%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12           12.25m ±  6%   12.39m ±  3%        ~ (p=0.818 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12          13.28m ±  0%   13.29m ±  2%        ~ (p=1.000 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12            6.042m ±  5%   5.409m ±  2%  -10.47% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12            7.824m ±  1%   7.418m ± 10%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12           9.432m ± 12%   8.988m ± 18%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12           20.09m ± 12%   18.80m ±  5%   -6.42% (p=0.026 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12           22.13m ±  6%   20.25m ±  3%   -8.49% (p=0.004 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12          23.64m ±  4%   21.90m ±  8%   -7.36% (p=0.015 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12         7.860m ±  3%   6.827m ±  2%  -13.15% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12        10.512m ±  5%   9.817m ±  1%   -6.61% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12        13.66m ±  6%   12.79m ±  6%   -6.37% (p=0.009 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12        29.11m ±  4%   28.72m ±  8%        ~ (p=0.240 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12        32.58m ±  3%   31.80m ±  2%   -2.39% (p=0.041 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12       35.14m ±  4%   34.43m ±  1%   -2.02% (p=0.026 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12              341.5n ±  1%   300.2n ±  1%  -12.08% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12              339.0n ±  2%   300.2n ±  1%  -11.43% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12             341.5n ±  1%   300.8n ±  6%  -11.95% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12             342.8n ±  2%   299.2n ±  1%  -12.69% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12             343.1n ± 18%   300.2n ±  1%  -12.48% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12            342.6n ±  1%   302.1n ±  4%  -11.84% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12           1.925m ±  3%   1.835m ±  4%   -4.66% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12           3.283m ±  5%   3.304m ±  7%        ~ (p=0.937 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12          4.385m ±  5%   4.367m ± 10%        ~ (p=0.818 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12          10.83m ±  1%   10.99m ±  5%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12          12.35m ±  3%   12.16m ±  2%        ~ (p=0.065 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12         13.71m ± 16%   13.33m ±  1%   -2.83% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12           5.990m ±  3%   5.463m ±  1%   -8.80% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12           8.023m ±  5%   7.482m ±  1%   -6.75% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12          9.542m ±  7%   8.779m ±  3%   -7.99% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12          19.22m ±  6%   18.11m ±  2%   -5.81% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12          21.80m ± 15%   20.04m ±  6%   -8.09% (p=0.009 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12         24.08m ± 12%   21.36m ±  9%  -11.28% (p=0.026 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12        7.682m ± 32%   6.980m ±  3%   -9.14% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12        10.69m ±  3%   10.26m ±  5%   -3.96% (p=0.026 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12       13.89m ± 10%   13.41m ±  4%        ~ (p=0.310 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12       29.62m ±  2%   28.89m ±  1%   -2.45% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12       34.73m ± 10%   32.56m ±  2%   -6.25% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12      36.32m ±  7%   36.00m ±  4%        ~ (p=0.310 n=6)
geomean                                                                                                                                         1.689m         1.593m         -5.70%

                                                                                                                                           │ frommetrics.txt │   prometheusconverter-frommetrics.txt    │
                                                                                                                                           │      B/op       │     B/op      vs base                    │
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12            1.499Mi ± 0%   1.383Mi ± 0%    -7.75% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12            3.611Mi ± 0%   3.495Mi ± 0%    -3.22% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12           5.730Mi ± 0%   5.612Mi ± 0%    -2.04% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12           7.022Mi ± 0%   6.905Mi ± 0%    -1.66% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12           9.161Mi ± 0%   9.042Mi ± 0%    -1.30% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12          11.30Mi ± 0%   11.19Mi ± 0%    -1.01% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12            6.825Mi ± 0%   6.354Mi ± 0%    -6.90% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12            9.695Mi ± 0%   9.223Mi ± 0%    -4.86% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12           12.67Mi ± 0%   12.20Mi ± 0%    -3.73% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12           20.52Mi ± 0%   20.05Mi ± 0%    -2.30% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12           23.41Mi ± 0%   22.93Mi ± 0%    -2.02% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12          26.39Mi ± 0%   25.92Mi ± 0%    -1.79% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12         8.062Mi ± 0%   7.574Mi ± 0%    -6.05% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12         13.06Mi ± 0%   12.57Mi ± 0%    -3.73% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12        18.18Mi ± 0%   17.69Mi ± 0%    -2.69% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12        27.42Mi ± 0%   26.93Mi ± 0%    -1.79% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-12        32.50Mi ± 0%   32.02Mi ± 0%    -1.50% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-12       37.72Mi ± 0%   37.24Mi ± 0%    -1.27% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-12                 48.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-12                 48.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-12                48.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-12                48.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-12                48.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-12               48.00 ± 0%      0.00 ± 0%  -100.00% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-12            1.499Mi ± 0%   1.383Mi ± 0%    -7.75% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-12            3.612Mi ± 0%   3.496Mi ± 0%    -3.22% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-12           5.730Mi ± 0%   5.613Mi ± 0%    -2.04% (p=0.002 n=6)
FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-12           7.024Mi ± 0%   6.906Mi ± 0%    -1.68% (p=0.002 n=6)
…